Polynomial multiplication
Using the FFT/IFFT technique, you are required to multiply long integers (of say 1024 digits).
Develop a suitable data structure for representing such long integers.

Populate such data structures, as required, to obtain such integers with random digits, the number of digits to generate should be taken as a parameter. That way smaller integers may be generated for testing and larger integers may be generated for experimentation.